requirement.txt
torch
torchvision
transformers
pandas
numpy
scikit-learn
Pillow
flask
kaggle